+++ packages/start-client-core/src/client-rpc/index.ts
@@ -1,1 +1,11 @@
 export { createClientRpc } from './createClientRpc'
+export {
+  FRAME_HEADER_SIZE,
+  FRAME_TYPE_CHUNK,
+  FRAME_TYPE_END,
+  FRAME_TYPE_ERROR,
+  FRAME_TYPE_JSON,
+  MAX_FRAMED_STREAMS,
+  MAX_FRAME_PAYLOAD_SIZE,
+  MAX_UNREAD_RAW_STREAM_BYTES,
+} from '../constants'